PSP-10834 : Finalize "add property to file logic"
