+++ source/frontend/src/features/mapSideBar/shared/update/properties/UpdateProperties.tsx
@@ -2,15 +2,18 @@ import axios, { AxiosError } from 'axios';
 import { FieldArray, Formik, FormikProps, getIn } from 'formik';
 import { geoJSON } from 'leaflet';
 import noop from 'lodash/noop';
-import { useContext, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useContext, useEffect, useMemo, useRef, useState } from 'react';
 import { Col, Row } from 'react-bootstrap';
 import { PiCornersOut } from 'react-icons/pi';
 import { toast } from 'react-toastify';
+import styled from 'styled-components';
 
+import { Button } from '@/components/common/buttons';
 import { LinkButton } from '@/components/common/buttons/LinkButton';
 import GenericModal from '@/components/common/GenericModal';
 import LoadingBackdrop from '@/components/common/LoadingBackdrop';
 import { useMapStateMachine } from '@/components/common/mapFSM/MapStateMachineContext';
+import { SelectedFeatureDataset } from '@/components/common/mapFSM/useLocationFeatureLoader';
 import { Section } from '@/components/common/Section/Section';
 import TooltipWrapper from '@/components/common/TooltipWrapper';
 import SelectedPropertyHeaderRow from '@/components/propertySelector/selectedPropertyList/SelectedPropertyHeaderRow';
@@ -22,10 +25,17 @@ import { useFeatureDatasetsWithAddresses } from '@/hooks/useFeatureDatasetsWithA
 import { getCancelModalProps, useModalContext } from '@/hooks/useModalContext';
 import { ApiGen_Concepts_File } from '@/models/api/generated/ApiGen_Concepts_File';
 import { UserOverrideCode } from '@/models/api/UserOverrideCode';
-import { arePropertyFormsEqual, exists, isValidId, latLngLiteralToGeometry } from '@/utils';
+import {
+  arePropertyFormsEqual,
+  exists,
+  firstOrNull,
+  isValidId,
+  latLngLiteralToGeometry,
+} from '@/utils';
 
 import { FileForm, PropertyForm } from '../../models';
 import SidebarFooter from '../../SidebarFooter';
+import AddPropertiesGuide from './AddPropertiesGuide';
 import { UpdatePropertiesYupSchema } from './UpdatePropertiesYupSchema';
 
 export interface IUpdatePropertiesProps {
@@ -53,7 +63,13 @@ export const UpdateProperties: React.FunctionComponent<IUpdatePropertiesProps> =
   const [isValid, setIsValid] = useState<boolean>(true);
   const { setModalContent, setDisplayModal } = useModalContext();
   const { resetFilePropertyLocations } = useContext(SideBarContext);
-  const { requestFlyToBounds, selectedFeatures, processCreation } = useMapStateMachine();
+  const {
+    requestFlyToBounds,
+    selectedFeatures,
+    processCreation,
+    mapLocationFeatureDataset,
+    prepareForCreation,
+  } = useMapStateMachine();
 
   useEditPropertiesMode();
 
@@ -166,6 +182,34 @@ export const UpdateProperties: React.FunctionComponent<IUpdatePropertiesProps> =
     }
   };
 
+  const selectedFeatureDataset = useMemo<SelectedFeatureDataset>(() => {
+    return {
+      selectingComponentId: mapLocationFeatureDataset?.selectingComponentId ?? null,
+      location: mapLocationFeatureDataset?.location,
+      fileLocation: mapLocationFeatureDataset?.fileLocation ?? null,
+      parcelFeature: firstOrNull(mapLocationFeatureDataset?.parcelFeatures),
+      pimsFeature: firstOrNull(mapLocationFeatureDataset?.pimsFeatures),
+      regionFeature: mapLocationFeatureDataset?.regionFeature ?? null,
+      districtFeature: mapLocationFeatureDataset?.districtFeature ?? null,
+      municipalityFeature: firstOrNull(mapLocationFeatureDataset?.municipalityFeatures),
+      isActive: true,
+      displayOrder: 0,
+    };
+  }, [
+    mapLocationFeatureDataset?.selectingComponentId,
+    mapLocationFeatureDataset?.location,
+    mapLocationFeatureDataset?.fileLocation,
+    mapLocationFeatureDataset?.parcelFeatures,
+    mapLocationFeatureDataset?.pimsFeatures,
+    mapLocationFeatureDataset?.regionFeature,
+    mapLocationFeatureDataset?.districtFeature,
+    mapLocationFeatureDataset?.municipalityFeatures,
+  ]);
+
+  const handleAddToSelection = useCallback(() => {
+    prepareForCreation([selectedFeatureDataset]);
+  }, [prepareForCreation, selectedFeatureDataset]);
+
   return (
     <>
       <LoadingBackdrop show={bcaLoading} />
@@ -181,60 +225,68 @@ export const UpdateProperties: React.FunctionComponent<IUpdatePropertiesProps> =
           />
         }
       >
-        <Formik<FileForm>
-          innerRef={formikRef}
-          initialValues={formFile}
-          validationSchema={UpdatePropertiesYupSchema}
-          onSubmit={async (values: FileForm) => {
-            const file: ApiGen_Concepts_File = values.toApi();
-            await saveFile(file);
-          }}
-        >
-          {formikProps => (
-            <FieldArray name="properties">
-              {({ remove }) => (
-                <Section
-                  header={
-                    <Row>
-                      <Col xs="11">Selected Properties</Col>
-                      <Col>
-                        <TooltipWrapper
-                          tooltip="Fit map to the file properties"
-                          tooltipId="property-selector-tooltip"
-                        >
-                          <LinkButton title="Fit boundaries button" onClick={fitBoundaries}>
-                            <PiCornersOut size={18} className="mr-2" />
-                          </LinkButton>
-                        </TooltipWrapper>
-                      </Col>
-                    </Row>
-                  }
-                >
-                  <SelectedPropertyHeaderRow />
-                  {formikProps.values.properties.map((property, index) => (
-                    <SelectedPropertyRow
-                      key={`property.${property.latitude}-${property.longitude}-${property.pid}-${property.apiId}`}
-                      onRemove={async () => {
-                        if (!property.apiId || (await props.canRemove(property.apiId))) {
-                          remove(index);
-                        } else {
-                          setShowAssociatedEntityWarning(true);
-                        }
-                      }}
-                      nameSpace={`properties.${index}`}
-                      index={index}
-                      property={property.toFeatureDataset()}
-                      showDisable={props.disableProperties}
-                    />
-                  ))}
-                  {formikProps.values.properties.length === 0 && (
-                    <span>No Properties selected</span>
-                  )}
-                </Section>
-              )}
-            </FieldArray>
+        <>
+          <AddPropertiesGuide />
+          {exists(selectedFeatureDataset?.parcelFeature) && (
+            <StyledButtonWrapper>
+              <Button onClick={handleAddToSelection}>Add selected property</Button>
+            </StyledButtonWrapper>
           )}
-        </Formik>
+          <Formik<FileForm>
+            innerRef={formikRef}
+            initialValues={formFile}
+            validationSchema={UpdatePropertiesYupSchema}
+            onSubmit={async (values: FileForm) => {
+              const file: ApiGen_Concepts_File = values.toApi();
+              await saveFile(file);
+            }}
+          >
+            {formikProps => (
+              <FieldArray name="properties">
+                {({ remove }) => (
+                  <Section
+                    header={
+                      <Row>
+                        <Col xs="11">Selected Properties</Col>
+                        <Col>
+                          <TooltipWrapper
+                            tooltip="Fit map to the file properties"
+                            tooltipId="property-selector-tooltip"
+                          >
+                            <LinkButton title="Fit boundaries button" onClick={fitBoundaries}>
+                              <PiCornersOut size={18} className="mr-2" />
+                            </LinkButton>
+                          </TooltipWrapper>
+                        </Col>
+                      </Row>
+                    }
+                  >
+                    <SelectedPropertyHeaderRow />
+                    {formikProps.values.properties.map((property, index) => (
+                      <SelectedPropertyRow
+                        key={`property.${property.latitude}-${property.longitude}-${property.pid}-${property.apiId}`}
+                        onRemove={async () => {
+                          if (!property.apiId || (await props.canRemove(property.apiId))) {
+                            remove(index);
+                          } else {
+                            setShowAssociatedEntityWarning(true);
+                          }
+                        }}
+                        nameSpace={`properties.${index}`}
+                        index={index}
+                        property={property.toFeatureDataset()}
+                        showDisable={props.disableProperties}
+                      />
+                    ))}
+                    {formikProps.values.properties.length === 0 && (
+                      <span>No Properties selected</span>
+                    )}
+                  </Section>
+                )}
+              </FieldArray>
+            )}
+          </Formik>
+        </>
       </MapSideBarLayout>
       <GenericModal
         variant="info"
@@ -306,3 +358,14 @@ export const addPropertiesToCurrentFile = <T extends { [key: string]: any }>(
   }
   notifyAddComplete();
 };
+
+const StyledButtonWrapper = styled.div`
+  margin: 0 1.6rem;
+  padding-left: 1.6rem;
+  text-align: left;
+  text-underline-offset: 2px;
+
+  button {
+    font-size: 14px;
+  }
+`;
